8_spring-docs/requirements.txt
streamlit
langchain
python-dotenv
pypdf
chromadb
replicate
Pillow